Revamped the way we run Jetty and GWT devmode.
We forgo the use of the stock Jetty built-into GWT devmode so that we can run a
pseudo-webapp at / that serves up our stock website images and CSS, and then we
run our actual webapp mounted at the place where it would normally be mounted.

This is simpler than requiring the user to have Apache and mod_jk and all of
the bits from a production webserver installed just so that they can test the
webapp in its proper context.

One downside is that if you browse the normal website, any use of Apache's <!--
in the webapp, but it would be neat if they did. Maybe I'll customize Jetty's
DefaultServlet to process those directives some day.